feat: 게임 결과내역 추가 / 게임 결과내역 삭제 / 테스트 추가

+++ src/main/java/com/example/balanceGame/dto/BoardDetailDto.java
@@ -19,8 +19,8 @@ public class BoardDetailDto {
     private String boardTitle;
     private String leftContent;
     private String rightContent;
-    private int leftCount;
-    private int rightCount;
+    private Long leftCount;
+    private Long rightCount;
     private Long heartCount;
     private List<CommentDto> comments;
 }
